mbe

Module expander

This module takes a (parsed) definition of macro_rules invocation, a tt::TokenTree representing an argument of macro invocation, and produces a tt::TokenTree for the result of the expansion.

Modules

matcher 🔒

An NFA-based parser, which is porting from rustc mbe parsing code

transcriber 🔒

Transcriber takes a template, like fn $ident() {}, a set of bindings like $ident => foo, interpolates variables in the template, to get fn foo() {}

Structs

Bindings 🔒

The actual algorithm for expansion is not too hard, but is pretty tricky. Bindings structure is the key to understanding what we are doing here.

Enums

Binding 🔒

Fragment 🔒

TokensOrigin 🔒

Functions

expand_rules 🔒