mbe::expander

Module matcher

An NFA-based parser, which is porting from rustc mbe parsing code

See https://github.com/rust-lang/rust/blob/70b18bc2cbac4712020019f5bf57c00905373205/compiler/rustc_expand/src/mbe/macro_parser.rs Here is a quick intro to how the parser works, copied from rustc:

A ‘position’ is a dot in the middle of a matcher, usually represented as a dot. For example · a $( a )* a b is a position, as is a $( · a )* a b.

The parser walks through the input a character at a time, maintaining a list of threads consistent with the current position in the input string: cur_items.

As it processes them, it fills up eof_items with threads that would be valid if the macro invocation is now over, bb_items with threads that are waiting on a Rust non-terminal like $e:expr, and next_items with threads that are waiting on a particular token. Most of the logic concerns moving the · through the repetitions indicated by Kleene stars. The rules for moving the · without consuming any input are called epsilon transitions. It only advances or calls out to the real Rust parser when no cur_items threads remain.

Example:

Start parsing a a a a b against [· a $( a )* a b].

Remaining input: a a a a b
next: [· a $( a )* a b]

- - - Advance over an a. - - -

Remaining input: a a a b
cur: [a · $( a )* a b]
Descend/Skip (first item).
next: [a $( · a )* a b]  [a $( a )* · a b].

- - - Advance over an a. - - -

Remaining input: a a b
cur: [a $( a · )* a b]  [a $( a )* a · b]
Follow epsilon transition: Finish/Repeat (first item)
next: [a $( a )* · a b]  [a $( · a )* a b]  [a $( a )* a · b]

- - - Advance over an a. - - - (this looks exactly like the last step)

Remaining input: a b
cur: [a $( a · )* a b]  [a $( a )* a · b]
Follow epsilon transition: Finish/Repeat (first item)
next: [a $( a )* · a b]  [a $( · a )* a b]  [a $( a )* a · b]

- - - Advance over an a. - - - (this looks exactly like the last step)

Remaining input: b
cur: [a $( a · )* a b]  [a $( a )* a · b]
Follow epsilon transition: Finish/Repeat (first item)
next: [a $( a )* · a b]  [a $( · a )* a b]  [a $( a )* a · b]

- - - Advance over a b. - - -

Remaining input: ''
eof: [a $( a )* a b ·]

Structs

BindingsBuilder 🔒

BindingsIdx 🔒

Match 🔒

MatchState 🔒

OpDelimitedIter 🔒

Enums

BindingKind 🔒

LinkNode 🔒

OpDelimited 🔒

Functions

collect_vars 🔒

eat_char 🔒

expect_lifetime 🔒

expect_separator 🔒

expect_tt 🔒

match_ 🔒

Matching errors are added to the Match.

match_loop 🔒

match_loop_inner 🔒

Process the matcher positions of cur_items until it is empty. In the process, this will produce more items in next_items, eof_items, and bb_items.

match_meta_var 🔒