mbe::expander::matcher

Function match_loop_inner

fn match_loop_inner<'t>(
    db: &dyn Database,
    src: TtIter<'t>,
    stack: &[TtIter<'t>],
    res: &mut Match<'t>,
    bindings_builder: &mut BindingsBuilder<'t>,
    cur_items: &mut SmallVec<[MatchState<'t>; 1]>,
    bb_items: &mut SmallVec<[MatchState<'t>; 1]>,
    next_items: &mut Vec<MatchState<'t>>,
    eof_items: &mut SmallVec<[MatchState<'t>; 1]>,
    error_items: &mut SmallVec<[MatchState<'t>; 1]>,
    delim_span: DelimSpan,
)

Process the matcher positions of cur_items until it is empty. In the process, this will produce more items in next_items, eof_items, and bb_items.

For more info about the how this happens, see the module-level doc comments and the inline comments of this function.

Parameters

src: the current token of the parser.

stack: the “parent” frames of the token tree

res: the match result to store errors

cur_items: the set of current items to be processed. This should be empty by the end of a successful execution of this function.

next_items: the set of newly generated items. These are used to replenish cur_items in the function parse.

eof_items: the set of items that would be valid if this was the EOF.

bb_items: the set of items that are waiting for the black-box parser.

error_items: the set of items in errors, used for error-resilient parsing