mbe::expander

Struct Bindings

struct Bindings<'a> {
    inner: FxHashMap<Symbol, Binding<'a>>,
}

The actual algorithm for expansion is not too hard, but is pretty tricky. Bindings structure is the key to understanding what we are doing here.

On the high level, it stores mapping from meta variables to the bits of syntax it should be substituted with. For example, if $e:expr is matched with 1 + 1 by macro_rules, the Binding will store $e -> 1 + 1.

The tricky bit is dealing with repetitions ($()*). Consider this example:

macro_rules! foo {
    ($($ i:ident $($ e:expr),*);*) => {
        $(fn $ i() { $($ e);*; })*
    }
}
foo! { foo 1,2,3; bar 4,5,6 }

Here, the $i meta variable is matched first with foo and then with bar, and $e is matched in turn with 1, 2, 3, 4, 5, 6.

To represent such โmulti-mappingsโ, we use a recursive structures: we map variables not to values, but to lists of values or other lists (that is, to the trees).

For the above example, the bindings would store

i -> [foo, bar]
e -> [[1, 2, 3], [4, 5, 6]]

We construct Bindings in the match_lhs. The interesting case is TokenTree::Repeat, where we use push_nested to create the desired nesting structure.

The other side of the puzzle is expand_subtree, where we use the bindings to substitute meta variables in the output template. When expanding, we maintain a nesting stack of indices which tells us which occurrence from the Bindings we should take. We push to the stack when we enter a repetition.

In other words, Bindings is a multi mapping from Symbol to tt::TokenTree, where the index to select a particular TokenTree among many is not a plain usize, but a &[usize].

Fields
inner: FxHashMap<Symbol, Binding<'a>>
Implementations

impl<'a> Bindings<'a>

fn push_optional(&mut self, name: Symbol)

fn push_empty(&mut self, name: Symbol)

fn bindings(&self) -> impl Iterator<Item = &Binding<'a>>

impl<'t> Bindings<'t>

fn get(&self, name: &Symbol, span: Span) -> Result<&Binding<'t>, ExpandError>

fn get_fragment( &self, name: &Symbol, span: Span, nesting: &mut [NestingState], marker: impl Fn(&mut Span), ) -> Result<Fragment<'t>, ExpandError>