mbe::expander

Module transcriber

Transcriber takes a template, like fn $ident() {}, a set of bindings like $ident => foo, interpolates variables in the template, to get fn foo() {}

Structs

ExpandCtx ğŸ”’

NestingState ğŸ”’

Functions

count ğŸ”’

Handles ${count(t, depth)}. our_depth is the recursion depth and count_depth is the depth defined by the metavar expression.

expand_repeat ğŸ”’

expand_subtree ğŸ”’

expand_subtree_with_delimiter ğŸ”’

expand_var ğŸ”’

fix_up_and_push_path_tt ğŸ”’

Inserts the path separator :: between an identifier and its following generic argument list, and then pushes into the buffer. See Fragment::Path for why we need this fixup.

transcribe ğŸ”’