Crate mbe

mbe (short for Macro By Example) crate contains code for handling macro_rules macros. It uses TokenTree (from tt package) as the interface, although it contains some code to bridge SyntaxNodes and TokenTrees as well!

The tests for this functionality live in another crate: hir_def::macro_expansion_tests::mbe.

Structs

DeclarativeMacro

This struct contains AST for a single macro_rules definition. What might be very confusing is that AST has almost exactly the same shape as tt::TokenTree, but there’s a crucial difference: in macro rules, $ident and $()* have special meaning (see Var and Repeat data structures)

Delimiter

DummyTestSpanMap

Punct

ValueResult

Enums

CountError

DelimiterKind

ExpandError

ParseError

TopEntryPoint

Parse the whole of the input as a given syntactic construct.

Constants

DUMMY

Traits

SpanMapper

Functions

parse_exprs_with_sep

Split token tree with separate expr: $($e:expr)SEP*

parse_to_token_tree

Convert a string to a TokenTree. The spans of the subtree will be anchored to the provided anchor with the given context.

parse_to_token_tree_static_span

Convert a string to a TokenTree. The passed span will be used for all spans of the produced subtree.

syntax_node_to_token_tree

Converts a syntax tree to a [tt::Subtree] using the provided span map to populate the subtree’s spans.

syntax_node_to_token_tree_modified

Converts a syntax tree to a [tt::Subtree] using the provided span map to populate the subtree’s spans. Additionally using the append and remove parameters, the additional tokens can be injected or hidden from the output.

token_tree_to_syntax_node

Converts a [tt::Subtree] back to a [SyntaxNode]. The produced SpanMap contains a mapping from the syntax nodes offsets to the subtree’s spans.

Type Aliases

ExpandResult