Crate mbe

mbe (short for Macro By Example) crate contains code for handling macro_rules macros. It uses TokenTree (from tt package) as the interface, although it contains some code to bridge SyntaxNodes and TokenTrees as well!

The tes for this functionality live in another crate: hir_def::macro_expansion_tests::mbe.

Structs

DeclarativeMacro

This struct contains AST for a single macro_rules definition. What might be very confusing is that AST has almost exactly the same shape as tt::TokenTree, but there’s a crucial difference: in macro rules, $ident and $()* have special meaning (see Var and Repeat data structures)

Delimiter

ExpandResult

Punct

Shift

TokenMap

Maps tt::TokenId to the relative range of the original token.

Enums

DelimiterKind

ExpandError

Origin

ParseError

ParserEntryPoint

rust-analyzer parser allows you to choose one of the possible entry points.

Functions

parse_exprs_with_sep

Split token tree with separate expr: $($e:expr)SEP*

parse_to_token_tree

Convert a string to a TokenTree

syntax_node_to_token_tree

Convert the syntax node to a TokenTree (what macro will consume).

syntax_node_to_token_tree_censored

Convert the syntax node to a TokenTree (what macro will consume) with the censored range excluded.

token_tree_to_syntax_node